% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_random_bits}
\alias{sample_random_bits}
\title{Sample random 32-bit integers fast}
\usage{
sample_random_bits(n)
}
\arguments{
\item{n}{Integer, number of random variables to draw}
}
\value{
integer vector of length n
}
\description{
Sample random 32-bit integers fast
}
